third_party/WebKit/LayoutTests/platform/mac/fast/workers/storage/test-authorizer-sync-expected.txt

CONSOLE ERROR: error vacuuming database (1546 disk I/O error)
Test the database authorizer.
Beginning write transaction:
SQLITE_CREATE_TABLE allowed.
SQLITE_CREATE_INDEX allowed.
SQLITE_CREATE_TEMP_TABLE allowed.
SQLITE_CREATE_TEMP_TRIGGER allowed.
SQLITE_CREATE_TEMP_VIEW allowed.
SQLITE_CREATE_TRIGGER allowed.
SQLITE_CREATE_VIEW allowed.
SQLITE_CREATE_VTABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_READ allowed.
SQLITE_SELECT allowed.
SQLITE_DELETE allowed.
SQLITE_INSERT allowed.
SQLITE_UPDATE allowed.
SQLITE_PRAGMA not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ALTER_TABLE allowed.
SQLITE_ALTER_TABLE allowed.
SQLITE_TRANSACTION not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ATTACH not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DETACH not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_REINDEX allowed.
SQLITE_ANALYZE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_INDEX allowed.
SQLITE_DROP_TEMP_TABLE allowed.
SQLITE_DROP_TEMP_TRIGGER allowed.
SQLITE_DROP_TEMP_VIEW allowed.
SQLITE_DROP_TRIGGER allowed.
SQLITE_DROP_VIEW allowed.
SQLITE_DROP_TABLE allowed.
Write transaction succeeded.

Beginning read transactions:
SQLITE_CREATE_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_TABLE allowed.
SQLITE_CREATE_INDEX not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_TEMP_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_TEMP_TRIGGER not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_TEMP_VIEW not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_TRIGGER not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_VIEW not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_VTABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_CREATE_INDEX allowed.
SQLITE_CREATE_TEMP_TABLE allowed.
SQLITE_CREATE_TEMP_TRIGGER allowed.
SQLITE_CREATE_TEMP_VIEW allowed.
SQLITE_CREATE_TRIGGER allowed.
SQLITE_CREATE_VIEW allowed.
SQLITE_CREATE_VTABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_READ allowed.
SQLITE_SELECT allowed.
SQLITE_DELETE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_INSERT not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_UPDATE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_PRAGMA not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ALTER_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ALTER_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_TRANSACTION not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ATTACH not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DETACH not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_REINDEX not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_ANALYZE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_INDEX not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_TEMP_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_TEMP_TRIGGER not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_TEMP_VIEW not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_TRIGGER not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_VIEW not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
SQLITE_DROP_TABLE not allowed: SyntaxError: Invalid or unauthorized statement; or the number of arguments did not match the number of ? placeholders. (12)
Read transactions succeeded.
